extract genesis builders to new testutil package
